+++ retrofit/src/main/java/retrofit2/HttpException.java
@@ -13,13 +13,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package retrofit2.adapter.rxjava;
-
-import retrofit2.Response;
+package retrofit2;
 
 /** Exception for an unexpected, non-2xx HTTP response. */
-public final class HttpException extends retrofit2.HttpException {
+public class HttpException extends Exception {
+  private final int code;
+  private final String message;
+  private final transient Response<?> response;
+
   public HttpException(Response<?> response) {
-    super(response);
+    super("HTTP " + response.code() + " " + response.message());
+    this.code = response.code();
+    this.message = response.message();
+    this.response = response;
+  }
+
+  /** HTTP status code. */
+  public int code() {
+    return code;
+  }
+
+  /** HTTP status message. */
+  public String message() {
+    return message;
+  }
+
+  /**
+   * The full HTTP response. This may be null if the exception was serialized.
+   */
+  public Response<?> response() {
+    return response;
   }
 }
